Moved WriteStatus, Fresh, and Streaming in to the net module.
